+++ R/er_meas_rank.R
@@ -137,6 +137,9 @@ er_meas_rank <- function(B_SOILTYPE_AGR, B_GWL_CLASS, A_P_SG, B_SLOPE_DEGREE, M_
   # merge all measures to the given fields
   dt <- as.data.table(merge.data.frame(dt, dt.meas.av, all = TRUE))
   
+    # only select measures at field level
+    dt <- dt[level == 'field']
+    
   # rank is zero when measures are not applicable given the crop type
   
     # columns with the Ecoregelingen ranks and reward
@@ -186,7 +189,7 @@ er_meas_rank <- function(B_SOILTYPE_AGR, B_GWL_CLASS, A_P_SG, B_SLOPE_DEGREE, M_
       fs1 <- paste0('f',sector)
       fs2 <- fs0[!fs0 %in% fs1]
       dt[,c(fs1) := 1]
-      dt[,c(fs2) := 0]
+      if(length(fs2) >= 1) { dt[,c(fs2) := 0] }
     
       # estimate whether sector allows applicability
       dt[, fsector := fdairy * dairy + farable * arable + ftree_nursery * tree_nursery + fbulbs * bulbs]